Move utility function due to circular import
